+++ src/main/java/com/redbeemedia/enigma/core/analytics/SilentAnalyticsReporter.java
@@ -30,7 +30,7 @@ public class SilentAnalyticsReporter implements IAnalyticsReporter {
     }
 
     @Override
-    public void playbackStarted(long offsetTime, String playMode, String mediaLocator, Long referenceTime, Integer bitrate, String programId) {
+    public void playbackStarted(long offsetTime, String playMode, String mediaLocator, Long referenceTime, Integer bitrate, String programId, Integer duration) {
 
     }