+++ src/test/java/com/redbeemedia/enigma/core/analytics/AnalyticsReportedEventTest.java
@@ -92,7 +92,7 @@ public class AnalyticsReportedEventTest<E extends IAnalyticsEventType> {
         } else if (eventType == AnalyticsEvents.PLAYER_READY) {
             return reporter -> reporter.playbackPlayerReady(72442L, "MockTech", "1.0");
         } else if (eventType == AnalyticsEvents.STARTED) {
-            return reporter -> reporter.playbackStarted(35737L, "test", "http://mock.example.com/media/mock", null, null, null);
+            return reporter -> reporter.playbackStarted(35737L, "test", "http://mock.example.com/media/mock", null, null, null,575778);
         } else if (eventType == AnalyticsEvents.PAUSED) {
             return reporter -> reporter.playbackPaused(37357L);
         } else if (eventType == AnalyticsEvents.RESUMED) {