+++ src/test/java/com/redbeemedia/enigma/core/player/InternalPlaybackSessionTest.java
@@ -453,26 +453,26 @@ public class InternalPlaybackSessionTest {
         };
         internalPlaybackSession.onStart(enigmaPlayer);
 
-        Assert.assertEquals("[Playback.Created][Playback.Started]", eventLog.toString());
+        Assert.assertEquals("[Playback.Created]", eventLog.toString());
 
         internalPlaybackSession.setSelectedVideoTrack(new MockVideoTrack(72634));
 
-        Assert.assertEquals("[Playback.Created][Playback.Started]", eventLog.toString());
+        Assert.assertEquals("[Playback.Created]", eventLog.toString());
 
         for(IEnigmaPlayerListener playerListener : playerListeners) {
             playerListener.onStateChanged(EnigmaPlayerState.LOADED, EnigmaPlayerState.PLAYING);
         }
 
-        Assert.assertEquals("[Playback.Created][Playback.Started]", eventLog.toString());
+        Assert.assertEquals("[Playback.Created][Playback.Started, 72634]", eventLog.toString());
 
         internalPlaybackSession.setSelectedVideoTrack(new MockVideoTrack(99352));
-        Assert.assertEquals("[Playback.Created][Playback.Started][Playback.BitrateChanged, 99352]", eventLog.toString());
+        Assert.assertEquals("[Playback.Created][Playback.Started, 72634][Playback.BitrateChanged, 99352]", eventLog.toString());
 
         internalPlaybackSession.setSelectedVideoTrack(new MockVideoTrack(99352));
-        Assert.assertEquals("[Playback.Created][Playback.Started][Playback.BitrateChanged, 99352]", eventLog.toString());
+        Assert.assertEquals("[Playback.Created][Playback.Started, 72634][Playback.BitrateChanged, 99352]", eventLog.toString());
 
         internalPlaybackSession.setSelectedVideoTrack(new MockVideoTrack(123));
-        Assert.assertEquals("[Playback.Created][Playback.Started][Playback.BitrateChanged, 99352][Playback.BitrateChanged, 123]", eventLog.toString());
+        Assert.assertEquals("[Playback.Created][Playback.Started, 72634][Playback.BitrateChanged, 99352][Playback.BitrateChanged, 123]", eventLog.toString());
     }