+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # CHANGELOG
+* `3.5.0 `Release - [3.5.0](#3.5.0)&nbsp; - &nbsp; 2023-02-16
 * `3.4.12`Release - [3.4.12](#3.4.12)&nbsp; - &nbsp; 2022-12-06
 * `3.4.11`Release - [3.4.11](#3.4.11)&nbsp; - &nbsp; 2022-11-22
 * `3.4.10`Release - [3.4.10](#3.4.10)&nbsp; - &nbsp; 2022-11-08
@@ -29,6 +30,19 @@
 * `3.1.8` Release - [3.1.8](#3.1.8) &nbsp; - &nbsp; 2020-12-11
 * `3.1.7` Release - [3.1.7](#3.1.7) &nbsp; - &nbsp; 2020-12-04
 * `3.1.6` Release - [3.1.6](#3.1.6) &nbsp; - &nbsp; 2020-11-13
+
+# <a name="3.5.0"></a> 3.5.0
+#### Features
+`
+DevEnigmaRiverAndroidCore
+DevEnigmaRiverAndroidCast
+DevEnigmaRiverAndroidInternalTestUtils
+`
+* EMP-18960 - Migrate SDKs from exposure API to event sink
+* EMP-19254 - Missing Average Bitrate in analytics
+* EMP-19362 - Missing video length in analytics
+* EMP-18650 - Upgraded Chromecast library
+
 
 # <a name="3.4.12"></a> 3.4.12
 #### Features